made log a simple logf

+++ src/Kernel/kernel.cpp
@@ -5,22 +5,22 @@
 void Kernel::init(){
     // Init Basic Memory Management
     heap.InitializeHeap(0x100000, 0x100000);
-    log("Kernel > init > Heap Initialized");
+    logf("Kernel > init > Heap Initialized");
 
     // Initialize the New TextRenderer to the VGA Output Buffer
     renderer.init();
     renderer.ClearScreen();
-    log("Kernel > init > TextRenderer Initialized");
+    logf("Kernel > init > TextRenderer Initialized");
 }
 
 // Main Kernel Function
 void Kernel::main() {
-    log("Kernel > main > Booted up");
+    logf("Kernel > main > Booted up");
 
     init();
     renderer.printf("[ %cKernel%c - %cMain%c ] Init Finished\n", FRONT_CYAN | BACK_BLACK, STD_COLOR, FRONT_GREEN | BACK_BLACK, STD_COLOR);
 
-    log("Kernel > main > Reached end of Kernel\n");
+    logf("Kernel > main > Reached end of Kernel\n");
     renderer.printf("[ %cKernel%c - %cMain%c ] %cWARNING:%c Reached end of Kernel\n", FRONT_CYAN | BACK_BLACK, STD_COLOR, FRONT_GREEN | BACK_BLACK, STD_COLOR, FRONT_RED | BACK_BLACK, STD_COLOR);
     return;
 }
